- [upd] 修改多行业下单逻辑

+++ api-mall/src/main/java/com/hailu/cloud/api/mall/module/multiindustry/dao/MultiIndustryOrderMapper.java
@@ -30,7 +30,7 @@ public interface MultiIndustryOrderMapper {
      *
      * @mbggenerated 2019-11-25
      */
-    MultiIndustryOrder selectByPrimaryKey(Integer id);
+    MultiIndustryOrder selectByPrimaryKey(Long id);
 
     /**
      *
